backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
groq
slowapi
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
